refactor(timeline): Move the decryption retrying into a separate struct
